+++ app/controllers/api/locale.php
@@ -2,6 +2,8 @@
 
 global $utopia, $register, $request, $response, $projectDB, $project, $user, $audit;
 
+use Appwrite\Database\Document;
+use Appwrite\Utopia\Response;
 use Utopia\App;
 use Utopia\Locale\Locale;
 use GeoIp2\Database\Reader;
@@ -42,7 +44,7 @@ $utopia->get('/v1/locale')
                 $output['continentCode'] = $record->continent->code;
                 $output['eu'] = (\in_array($record->country->isoCode, $eu)) ? true : false;
 
-                foreach ($currencies as $code => $element) {
+                foreach ($currencies as $element) {
                     if (isset($element['locations']) && isset($element['code']) && \in_array($record->country->isoCode, $element['locations'])) {
                         $currency = $element['code'];
                     }
@@ -61,7 +63,9 @@ $utopia->get('/v1/locale')
             $response
                 ->addHeader('Cache-Control', 'public, max-age='.$time)
                 ->addHeader('Expires', \date('D, d M Y H:i:s', \time() + $time).' GMT') // 45 days cache
-                ->json($output);
+            ;
+
+            $response->dynamic(new Document($output), Response::MODEL_LOCALE);
         }
     );