+++ app/controllers/api/teams.php
@@ -3,7 +3,6 @@
 global $utopia, $register, $request, $response, $projectDB, $project, $user, $audit, $mail, $mode, $clients;
 
 use Utopia\Exception;
-use Utopia\Response;
 use Utopia\Config\Config;
 use Utopia\Validator\Email;
 use Utopia\Validator\Text;
@@ -19,6 +18,7 @@ use Appwrite\Database\Validator\UID;
 use Appwrite\Database\Validator\Authorization;
 use Appwrite\Database\Exception\Duplicate;
 use Appwrite\Template\Template;
+use Appwrite\Utopia\Response;
 
 $utopia->post('/v1/teams')
     ->desc('Create Team')
@@ -77,10 +77,8 @@ $utopia->post('/v1/teams')
                 }
             }
 
-            $response
-                ->setStatusCode(Response::STATUS_CODE_CREATED)
-                ->json($team->getArrayCopy())
-            ;
+            $response->setStatusCode(Response::STATUS_CODE_CREATED);
+            $response->dynamic($team, Response::MODEL_TEAM);
         }
     );
 
@@ -110,7 +108,10 @@ $utopia->get('/v1/teams')
                 ],
             ]);
 
-            $response->json(['sum' => $projectDB->getSum(), 'teams' => $results]);
+            $response->dynamic(new Document([
+                'sum' => $projectDB->getSum(),
+                'teams' => $results
+            ]), Response::MODEL_TEAM_LIST);
         }
     );
 
@@ -131,7 +132,7 @@ $utopia->get('/v1/teams/:teamId')
                 throw new Exception('Team not found', 404);
             }
 
-            $response->json($team->getArrayCopy([]));
+            $response->dynamic($team, Response::MODEL_TEAM);
         }
     );
 
@@ -161,7 +162,7 @@ $utopia->put('/v1/teams/:teamId')
                 throw new Exception('Failed saving team to DB', 500);
             }
 
-            $response->json($team->getArrayCopy());
+            $response->dynamic($team, Response::MODEL_TEAM);
         }
     );
 
@@ -364,21 +365,12 @@ $utopia->post('/v1/teams/:teamId/memberships')
                 ->setParam('resource', 'teams/'.$teamId)
             ;
 
-            $response
-                ->setStatusCode(Response::STATUS_CODE_CREATED) // TODO change response of this endpoint
-                ->json(\array_merge($membership->getArrayCopy([
-                    '$id',
-                    'userId',
-                    'teamId',
-                    'roles',
-                    'invited',
-                    'joined',
-                    'confirm',
-                ]), [
-                    'email' => $email,
-                    'name' => $name,
-                ]))
-            ;
+            $response->setStatusCode(Response::STATUS_CODE_CREATED); // TODO change response of this endpoint
+
+            $response->dynamic(new Document(\array_merge($membership->getArrayCopy(), [
+                'email' => $email,
+                'name' => $name,
+            ])), Response::MODEL_MEMBERSHIP);
         }
     );
 
@@ -425,18 +417,10 @@ $utopia->get('/v1/teams/:teamId/memberships')
 
                 $temp = $projectDB->getDocument($membership->getAttribute('userId', null))->getArrayCopy(['email', 'name']);
 
-                $users[] = \array_merge($temp, $membership->getArrayCopy([
-                    '$id',
-                    'userId',
-                    'teamId',
-                    'roles',
-                    'invited',
-                    'joined',
-                    'confirm',
-                ]));
+                $users[] = new Document(\array_merge($temp, $membership->getArrayCopy()));
             }
 
-            $response->json(['sum' => $projectDB->getSum(), 'memberships' => $users]);
+            $response->dynamic(new Document(['sum' => $projectDB->getSum(), 'memberships' => $users]), Response::MODEL_MEMBERSHIP_LIST);
         }
     );
 
@@ -557,19 +541,12 @@ $utopia->patch('/v1/teams/:teamId/memberships/:inviteId/status')
             $response
                 ->addCookie(Auth::$cookieName.'_legacy', Auth::encodeSession($user->getId(), $secret), $expiry, '/', COOKIE_DOMAIN, ('https' == $protocol), true, null)
                 ->addCookie(Auth::$cookieName, Auth::encodeSession($user->getId(), $secret), $expiry, '/', COOKIE_DOMAIN, ('https' == $protocol), true, COOKIE_SAMESITE)
-                ->json(\array_merge($membership->getArrayCopy([
-                    '$id',
-                    'userId',
-                    'teamId',
-                    'roles',
-                    'invited',
-                    'joined',
-                    'confirm',
-                ]), [
-                    'email' => $user->getAttribute('email'),
-                    'name' => $user->getAttribute('name'),
-                ]))
             ;
+
+            $response->dynamic(new Document(\array_merge($membership->getArrayCopy(), [
+                'email' => $user->getAttribute('email'),
+                'name' => $user->getAttribute('name'),
+            ])), Response::MODEL_MEMBERSHIP);
         }
     );